adlv1 refactoring

+++ src/c/include/common/types.h
@@ -2,19 +2,21 @@
 #define ADL_COMMON_TYPES_H
 
 
+//#define ADL_SYS_TYPES_H
+
 #ifndef ADL_SYS_TYPES_H
 
 
 typedef unsigned char      u8;
 typedef unsigned short     u16;
 typedef unsigned int       u32;
-typedef unsigned long      u64;
+typedef unsigned long long      u64;
 
 
 typedef signed char        s8;
 typedef signed short       s16;
 typedef signed int         s32;
-typedef signed long        s64;
+typedef signed long long        s64;
 
 
 #else
